V2/requirements.txt
python-dotenv
langchain
langchain_community
openai
slack-sdk
slack-bolt
flask